rag_001_naive_prompt/requirements.txt
faiss-cpu
openai
dotenv